+++ libtest/src/main/java/com/zalo/zing/migrate/MigratorAdapter.java
@@ -1,0 +1,10 @@
+package com.zalo.zing.migrate;
+
+import com.zing.zalo.zarcel.ZarcelMigrator;
+
+public class MigratorAdapter implements ZarcelMigrator<MigratorClass> {
+    @Override
+    public void migrate(MigratorClass object, int fromVersion, int toVersion) {
+        object.color = Integer.MAX_VALUE;
+    }
+}