+++ zarcel-annotations/src/main/java/com/zing/zalo/zarcel/ZarcelMigrator.java
@@ -1,0 +1,7 @@
+package com.zing.zalo.zarcel;
+
+import com.zing.zalo.data.serialization.Serializable;
+
+public interface ZarcelMigrator<T extends Serializable> {
+    void migrate(T object, int fromVersion, int toVersion);
+}